pyproject.toml
[tool.poetry]
name = "credit-card"
version = "0.2.3"
description = "credit card api"
authors = ["Andrey <megagran.work@gmail.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
fastapi = "^0.103.0"
uvicorn = "^0.23.2"
pyjwt = "^2.8.0"
python-multipart = "^0.0.6"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
httpx = "^0.24.1"
pytest-asyncio = "^0.21.1"
pytest-mock = "^3.11.1"
pydantic-settings = "^2.0.3"
aiokafka = "^0.8.1"
aiohttp = "^3.8.5"
alembic = "^1.12.0"
sqlalchemy = "^2.0.21"
asyncpg = "^0.28.0"

[tool.poetry.group.ml.dependencies]
h5py = "^3"
scipy = "^1.11"
tensorflow-io = {extras = ["tensorflow"], version = "^0.27"}
deepface = "^0.0.79"

[tool.poetry.dev-dependencies]
pre-commit = "^3.3.3"
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
mypy = "^1.5.1"
wemake-python-styleguide = "^0.18.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
